virtualenv note

+++ bootstrap.sh
@@ -7,5 +7,9 @@ git submodule init || true
 ## work (as the required commits are usually not within that depth)
 git submodule update --recursive
 ./0dotlink
-test -e ./.emacs.d/.python-environments/default/bin/pip && \
- ./.emacs.d/.python-environments/default/bin/pip install jedi epc flake8 pylint
+if [ -e ./.emacs.d/.python-environments/default/bin/pip ]; then
+    echo "Attempting to install required python-modules"
+    ./.emacs.d/.python-environments/default/bin/pip install jedi epc flake8 pylint
+else
+    echo "WARN: did not find virtualenv at ../../.virtualenv"
+fi
